Update utils/handlePlugin (rename, add functions)
