+++ packages/bunadmin/src/utils/scripts/index.ts
@@ -1,7 +1,11 @@
 import { ENV } from "@/utils"
 
 export { default as dataToGql } from "./dataToGql"
-export { default as handleSpecialPlugin } from "./handleSpecialPlugin"
+export {
+  handlePluginPath,
+  specialPluginGroup,
+  specialPluginSlug
+} from "./handlePlugin"
 export { default as initData } from "./initData"
 export { fixTreeDataTr } from "./muiTable"
 export { default as request } from "./request"
